requirements.txt
Package            Version
------------------ -----------
annotated-types    0.7.0
attrs              25.1.0
cachetools         5.5.1
certifi            2025.1.31
cffi               1.17.1
charset-normalizer 3.4.1
click              8.1.8
colorama           0.4.6
google-auth        2.38.0
google-genai       1.0.0
h11                0.14.0
idna               3.10
joblib             1.4.2
nltk               3.9.1
outcome            1.3.0.post0
pip                24.3.1
pyasn1             0.6.1
pyasn1_modules     0.4.1
pycparser          2.22
pydantic           2.10.6
pydantic_core      2.27.2
PySocks            1.7.1
python-dotenv      1.0.1
regex              2024.11.6
requests           2.32.3
rsa                4.9
selenium           4.28.1
sniffio            1.3.1
sortedcontainers   2.4.0
textblob           0.19.0
tqdm               4.67.1
trio               0.28.0
trio-websocket     0.11.1
typing_extensions  4.12.2
urllib3            2.3.0
websocket-client   1.8.0
websockets         14.2
wsproto            1.2.0
